Renamed dev/lib/ to dev/utils/ to match project practices.
