+++ dev/tasks/dev/tasks/status.js
@@ -5,10 +5,10 @@
 
 'use strict';
 
-const tools = require( '../../../lib/tools' );
-const git = require( '../../../lib/git' );
+const tools = require( '../../../utils/tools' );
+const git = require( '../../../utils/git' );
 const path = require( 'path' );
-const log = require( '../../../lib/log' );
+const log = require( '../../../utils/log' );
 
 /**
  * 1. Get CKEditor5 dependencies from package.json file.
