+++ dev/tasks/dev/tasks/update.js
@@ -5,10 +5,10 @@
 
 'use strict';
 
-const tools = require( '../../../lib/tools' );
-const git = require( '../../../lib/git' );
+const tools = require( '../../../utils/tools' );
+const git = require( '../../../utils/git' );
 const path = require( 'path' );
-const log = require( '../../../lib/log' );
+const log = require( '../../../utils/log' );
 
 /**
  * 1. Fetch all branches from each origin in main CKEditor 5 repository.
