+++ dev/tests/test/tasks.js
@@ -9,7 +9,7 @@
 
 const tasks = require( '../../tasks/test/tasks' )();
 const buildUtils = require( '../../tasks/build/utils' );
-const devTools = require( '../../lib/tools' );
+const devTools = require( '../../utils/tools' );
 const Vinyl = require( 'vinyl' );
 
 describe( 'test-node', () => {
